third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-special-urls.html

<!doctype html>
<title>Credential Manager: Call get() across browsing contexts.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/credential-helpers.js"></script>
<body>
<script>

var PROBE_CREDENTIALS = "window.parent.postMessage(String(navigator.credentials), '*');";
var PROBE_CREDENTIALS_SCRIPT = "<script>" + PROBE_CREDENTIALS + "</scr" + "ipt>";

promise_test(t => {
  let frame = document.createElement("iframe");
  frame.src = "data:text/html," + PROBE_CREDENTIALS_SCRIPT;
  window.setTimeout(_ => document.body.append(frame));

  let eventWatcher = new EventWatcher(t, window, "message");
  return eventWatcher.wait_for("message").then(message => {
    assert_equals(message.data, "undefined");
  });
}, "data: URL");


// ----------------------

var TEST_NESTED_CREDENTIAL_ID = "nestedCredentialId";

var GET_CREDENTIALS = "navigator.credentials.get({}).then(c => window.parent.postMessage(c.id, '*'));";
var GET_CREDENTIALS_SCRIPT = "<script>" + GET_CREDENTIALS + "</scr" + "ipt>";

 promise_test(t => {
   let frame = document.createElement("iframe");
   frame.src = "resources/nested_credential_manager_client.html";
   window.setTimeout(_ => document.body.append(frame));
 
   let loadWatcher = new EventWatcher(t, frame, "load");
   loadWatcher.wait_for("load").then(_ =>
       frame.contentWindow.location = "javascript:" + GET_CREDENTIALS);
 
   let messageWatcher = new EventWatcher(t, window, "message");
   return messageWatcher.wait_for("message").then(message => {
     assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
   });
 }, "javascript:");

// -------------------------

var templateFrame = document.createElement("iframe");
templateFrame.src = "resources/nested_credential_manager_client.html";
templateFrame.addEventListener("load", _ => {

   promise_test(t => {
     let frame = document.createElement("iframe");
     frame.srcdoc = templateFrame.contentDocument.documentElement.outerHTML;
     window.setTimeout(_ => document.body.append(frame));
 
     let loadWatcher = new EventWatcher(t, frame, "load");
     loadWatcher.wait_for("load").then(_ => {
       frame.contentWindow.eval(GET_CREDENTIALS);
     });
 
     let eventWatcher = new EventWatcher(t, window, "message");
     return eventWatcher.wait_for("message").then(message => {
       assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
     });
   }, "srcdoc");
 
   promise_test(t => {
     let frame = document.createElement("iframe");
     frame.src = "about:blank";
     window.setTimeout(_ => document.body.append(frame));
 
     let loadWatcher = new EventWatcher(t, frame, "load");
     loadWatcher.wait_for("load").then(_ => {
       frame.contentDocument.write(templateFrame.contentDocument.documentElement.outerHTML);
       frame.contentDocument.write(GET_CREDENTIALS_SCRIPT);
     });
 
     let eventWatcher = new EventWatcher(t, window, "message");
     return eventWatcher.wait_for("message").then(message => {
       assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
     });
   }, "blank v1");

  promise_test(t => {
    let frame = document.createElement("iframe");
    frame.src = "about:blank";
    window.setTimeout(_ => document.body.append(frame));

    let loadWatcher = new EventWatcher(t, frame, "load");
    loadWatcher.wait_for("load").then(_ => {
      frame.contentWindow.mojo = {config: {autoLoadMojomDeps: false}};
      var lastScript = null;
      for (let script of templateFrame.contentDocument.head.getElementsByTagName("script")) {
        if (!script.src)
          continue;
        let copy = document.createElement("script");
        copy.src = script.src;
        copy.async = false;
        lastScript = copy;
        window.setTimeout(_ => frame.contentDocument.body.append(copy));
      }

      let lastScriptLoadedWatcher = new EventWatcher(t, lastScript, "load");
      lastScriptLoadedWatcher.wait_for("load").then(_ => {
        frame.contentWindow.mockCredentialManager.setResponse(TEST_NESTED_CREDENTIAL_ID, "p", "", "");
        frame.contentWindow.eval(GET_CREDENTIALS);
      });
    });

    let eventWatcher = new EventWatcher(t, window, "message");
    return eventWatcher.wait_for("message").then(message => {
      assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
    });
  }, "blank");
});
document.body.append(templateFrame);

</script>
